Create two files, login.jsp and profile.jsp.
When a user visits /login, they should see a form for logging in

This form should have a 'username' and 'password' field.

This form should submit a POST request to /login

If the username submitted is "admin", and the password is "password", redirect the user to the profile page, otherwise, redirect back to the login form.